+++ node/src/database.rs
@@ -115,6 +115,7 @@ pub trait Ledger {
     fn faults_by_block(&self, start_height: u64) -> Result<Vec<Fault>>;
     fn faults(&self, faults_ids: &[[u8; 32]]) -> Result<Vec<Fault>>;
 
+    fn latest_block_opt(&self) -> Result<Option<LightBlock>>;
     fn latest_block(&self) -> Result<LightBlock>;
 }